+++ src/layers/GDDLLoginLayer.h
@@ -15,8 +15,9 @@ class GDDLLoginLayer final : public FLAlertLayer {
     CCTextInputNode* passwordTextField = nullptr;
     CCMenuItemSpriteExtra* loginButton = nullptr;
 
-    const inline static std::string loginEndpoint = time(nullptr) < Utils::API_SWITCH_TIME ? "https://gdladder.com/api/login" : "https://gdladder.com/api/account/login";
-    EventListener<web::WebTask> loginListener;
+    const inline static std::string loginEndpoint = "https://gdladder.com/api/account/login";
+    const inline static std::string meEndpoint = "https://gdladder.com/api/user/me";
+    EventListener<web::WebTask> loginListener, meListener;
     LoginSettingNodeV3* settingNode;
 
     bool init() override;
@@ -24,14 +25,14 @@ class GDDLLoginLayer final : public FLAlertLayer {
     void onLoginClicked(cocos2d::CCObject *sender);
 
     void prepareSearchListener();
-    void saveLoginData(const std::string& sid, const std::string& sig, int uid);
+    void prepareMeListener();
+    void saveLoginData(const std::string& sid, int uid);
     void closeLoginPanel();
     void showLoadingCircle();
     void hideLoadingCircle();
 
     // getting around geode::web limitations
     std::pair<std::string, std::string> getCookieValue(const char* content);
-    static size_t writeCallback(char *contents, size_t size, size_t nmemb, void *userp);
 
 public:
     static GDDLLoginLayer* create();