+++ src/layers/GDDLRatingSubmissionLayer.cpp
@@ -212,7 +212,7 @@ void GDDLRatingSubmissionLayer::onToggleSoloCompletion(CCObject* sender) {
  *  "rating":2,
  *  "enjoyment":6,
  *  "refreshRate":60,
- *  "device":1, - 1 for pc, 2 for mobile
+ *  "device": "pc" or "mobile" - api V2 change
  *  "proof":"https://youtu.be/3-BUEoH9WBs",
  *  "progress":100,
  *  "attempts":403,
@@ -221,7 +221,7 @@ void GDDLRatingSubmissionLayer::onToggleSoloCompletion(CCObject* sender) {
  *  }
  *
  *  headers:
- *  Cookie: gddl.sid.sig=<sid.sig>; gddl.sid=<sid>
+ *  Cookie: gddl.sid=<sid>
  */
 
 void GDDLRatingSubmissionLayer::onSubmitClicked(CCObject* sender) {
@@ -237,7 +237,7 @@ void GDDLRatingSubmissionLayer::onSubmitClicked(CCObject* sender) {
             // a request to retrieve the userid has to be made first before making the submission request
             std::string requestURL = userSearchEndpoint;
             requestedUsername = secondPlayerTextfield->getString();
-            requestURL += "?name=" + requestedUsername + (time(nullptr) < Utils::API_SWITCH_TIME ? "&chunk=25" : "&limit=25");
+            requestURL += "?name=" + requestedUsername + "&limit=25";
             auto req = web::WebRequest();
             req.header("User-Agent", Utils::getUserAgent());
             userSearchListener.setFilter(req.get(requestURL));
@@ -338,9 +338,9 @@ void GDDLRatingSubmissionLayer::prepareSubmissionListeners() {
     submissionListener.bind([this](web::WebTask::Event* e) {
         if (web::WebResponse* res = e->getValue()) {
             const auto jsonResponse = res->json().unwrapOr(matjson::Value());
-            if (res->code() == 200) {
+            if (res->code() == 201) {
                 std::string message = "Rating submitted!";
-                if (jsonResponse.contains("wasAuto") && jsonResponse["wasAuto"].asBool().unwrap()) {
+                if (jsonResponse.contains("wasAuto") && jsonResponse["wasAuto"].isBool() && jsonResponse["wasAuto"].asBool().unwrap()) {
                     message = "Submission accepted!";
                 }
                 // cache submitted submission
@@ -349,25 +349,21 @@ void GDDLRatingSubmissionLayer::prepareSubmissionListeners() {
                 Notification::create(message, NotificationIcon::Success, 2)->show();
                 onClose(nullptr);
             } else {
-                std::string error = "Unknown error";
-                if(jsonResponse.contains("error")) {
-                    error = jsonResponse["error"].asString().unwrap();
-                    if (error == "Authentication failed!") {
-                        LoginSettingNodeV3::logOut();
-                        GDDLLoginLayer::create()->show();
-                        Notification::create("Your session expired, log in again", NotificationIcon::Warning, 2)->show();
-                        return;
-                    }
+                const std::string error = jsonResponse["message"].asString().unwrapOr("Error while submitting rating - unknown error");
+                if (error == "Authentication failed!" || error == "Unauthorized") {
+                    LoginSettingNodeV3::logOut();
+                    GDDLLoginLayer::create()->show();
+                    Notification::create("Your session expired, log in again", NotificationIcon::Warning, 2)->show();
+                    return;
                 }
                 // cache that no submission was made
                 RatingsManager::cacheSubmission(this->gddlLevelID, Submission());
                 Notification::create(error, NotificationIcon::Error, 2)->show();
             }
-        }
-        else if (e->isCancelled()) {
+        } else if (e->isCancelled()) {
             // cache that no submission was made
             RatingsManager::cacheSubmission(this->gddlLevelID, Submission());
-            Notification::create("An error occurred", NotificationIcon::Error, 2)->show();
+            Notification::create("Error while submitting rating - request cancelled", NotificationIcon::Error, 2)->show();
         }
     });
     userSearchListener.bind([this](web::WebTask::Event* e) {
@@ -379,18 +375,15 @@ void GDDLRatingSubmissionLayer::prepareSubmissionListeners() {
                     submissionJson["secondPlayerID"] = id;
                     makeSubmissionRequest();
                 } else {
-                    Notification::create(id == -1 ? "Second player not found!" : "An error occurred", NotificationIcon::Error, 2)->show();
+                    Notification::create(id == -1 ? "Second player not found!" : "Error while finding second player - server returned invalid response", NotificationIcon::Error, 2)->show();
                 }
             } else {
-                std::string error = "Unknown error";
-                if(jsonResponse.contains("error")) {
-                    error = jsonResponse["error"].asString().unwrap();
-                }
+                const std::string error = jsonResponse["message"].asString().unwrapOr("Error while finding second player - unknown error");
                 Notification::create(error, NotificationIcon::Error, 2)->show();
             }
         }
         else if (e->isCancelled()) {
-            Notification::create("An error occurred", NotificationIcon::Error, 2)->show();
+            Notification::create("Error while finding second player - request cancelled", NotificationIcon::Error, 2)->show();
         }
     });
     userSubmissionCheckListener.bind([this](web::WebTask::Event* e) {
@@ -402,16 +395,17 @@ void GDDLRatingSubmissionLayer::prepareSubmissionListeners() {
                 RatingsManager::cacheSubmission(this->gddlLevelID, submission);
                 showAlreadySubmittedWarning();
             } else {
-                if (!jsonResponse.contains("error") || jsonResponse["error"].asString().unwrap() != "Submission not found!") {
-                    Notification::create("Check if already submitted failed", NotificationIcon::Warning, 2)->show();
-                } else {
+                const std::string error = jsonResponse["message"].asString().unwrapOr("Error while checking for existing submission - unknown error");
+                if (error == "Submission not found") {
                     // save an empty one
                     RatingsManager::cacheSubmission(this->gddlLevelID, Submission());
+                } else {
+                    Notification::create(error, NotificationIcon::Warning, 2)->show();
                 }
             }
         }
         else if (e->isCancelled()) {
-            Notification::create("Check if already submitted failed", NotificationIcon::Warning, 2)->show();
+            Notification::create("Error while checking for existing submission - request cancelled", NotificationIcon::Warning, 2)->show();
         }
     });
 }
@@ -425,11 +419,7 @@ bool GDDLRatingSubmissionLayer::isValidProof(const std::string& proofURL) {
 
 std::string GDDLRatingSubmissionLayer::fillOutSubmissionJson() {
     submissionJson["levelID"] = this->gddlLevelID;
-    if (time(nullptr) < Utils::API_SWITCH_TIME) {
-        submissionJson["device"] = mobile ? 2 : 1;
-    } else {
-        submissionJson["device"] = mobile ? "mobile" : "pc";
-    }
+    submissionJson["device"] = mobile ? "mobile" : "pc";
     const int correctedRating = std::min(std::max(0, Utils::getNumberTextfieldValue(ratingTextfield)), 35);
     if (correctedRating != 0) {
         submissionJson["rating"] = correctedRating;
@@ -490,8 +480,7 @@ void GDDLRatingSubmissionLayer::makeSubmissionRequest() {
     auto req = web::WebRequest();
     req.header("User-Agent", Utils::getUserAgent());
     req.bodyJSON(submissionJson);
-    req.header("Cookie", fmt::format("gddl.sid.sig={}; gddl.sid={}",
-                                     Mod::get()->getSavedValue<std::string>("login-sig", ""),
+    req.header("Cookie", fmt::format("gddl.sid={}",
                                      Mod::get()->getSavedValue<std::string>("login-sid", "")));
     submissionListener.setFilter(req.post(submissionEndpoint));
 }