+++ src/layers/GDDLSearchLayer.cpp
@@ -269,7 +269,7 @@ void GDDLSearchLayer::showPage() {
 
 void GDDLSearchLayer::loadValues() {
     if(!simplified) {
-        page = 1;
+        page = 0;
         nameTextfield->setString(name.c_str());
         Utils::setNumberWithDefZeroTextfield(lowTier, tierLowTextfield);
         Utils::setNumberWithDefZeroTextfield(highTier, tierHighTextfield);
@@ -326,7 +326,7 @@ void GDDLSearchLayer::saveValues() {
 
 // ReSharper disable once CppDFAUnreachableFunctionCall
 void GDDLSearchLayer::resetValues() {
-    page = 1;
+    page = 0;
     nameTextfield->setString("");
     Utils::setNumberWithDefZeroTextfield(0, tierLowTextfield);
     Utils::setNumberWithDefZeroTextfield(0, tierHighTextfield);
@@ -481,26 +481,23 @@ std::string GDDLSearchLayer::addBoolToRequest(const std::string &paramName, cons
 
 std::string GDDLSearchLayer::formSearchRequest() {
     std::string request = searchEndpoint;
-    request += "?page=" + std::to_string(onlinePagesFetched + 1) + (time(nullptr) < Utils::API_SWITCH_TIME ? "&chunk=50" : "&limit=25");
+    request += "?page=" + std::to_string(onlinePagesFetched) + "&limit=25";
     request += addStringToRequest("name", name);
-    request += addValueToRequest(time(nullptr) < Utils::API_SWITCH_TIME ? "lowTier" : "minRating", lowTier, 0);
-    request += addValueToRequest(time(nullptr) < Utils::API_SWITCH_TIME ? "highTier" : "maxRating", highTier, 0);
+    request += addValueToRequest("minRating", lowTier, 0);
+    request += addValueToRequest("maxRating", highTier, 0);
     request += addValueToRequest("difficulty", difficulty+1, 6); // API 1.9.0 - diffs 1-5
     request += addStringToRequest("creator", creator);
     request += addStringToRequest("song", song);
-    if (time(nullptr) < Utils::API_SWITCH_TIME) {
-        request += addBoolToRequest("exactName", exactName);
-    }
-    request += addBoolToRequest(time(nullptr) < Utils::API_SWITCH_TIME ? "removeUnrated" : "excludeUnrated", removeUnrated);
-    request += addBoolToRequest(time(nullptr) < Utils::API_SWITCH_TIME ? "removeUnratedEnj" : "excludeUnratedEnjoyment", removeUnratedEnj);
-    request += addBoolToRequest(time(nullptr) < Utils::API_SWITCH_TIME ? "removeRated" : "excludeRated", removeRated);
-    request += addBoolToRequest(time(nullptr) < Utils::API_SWITCH_TIME ? "removeRatedEnj" : "excludeRatedEnjoyment", removeRatedEnj);
-    request += addValueToRequest(time(nullptr) < Utils::API_SWITCH_TIME ? "subLowCount" : "minSubmissionCount", subLowCount, 0);
-    request += addValueToRequest(time(nullptr) < Utils::API_SWITCH_TIME ? "subHighCount" : "maxSubmissionCount", subHighCount, 0);
-    request += addValueToRequest(time(nullptr) < Utils::API_SWITCH_TIME ? "enjLowCount" : "minEnjoymentCount", enjLowCount, 0);
-    request += addValueToRequest(time(nullptr) < Utils::API_SWITCH_TIME ? "enjHighCount" : "maxEnjoymentCount", enjHighCount, 0);
-    request += addValueToRequest(time(nullptr) < Utils::API_SWITCH_TIME ? "enjLow" : "minEnjoyment", enjLow, 0.0f);
-    request += addValueToRequest(time(nullptr) < Utils::API_SWITCH_TIME ? "enjHigh" : "maxEnjoyment", enjHigh, highestEnjoyment);
+    request += addBoolToRequest("excludeUnrated", removeUnrated);
+    request += addBoolToRequest("excludeUnratedEnjoyment", removeUnratedEnj);
+    request += addBoolToRequest("excludeRated", removeRated);
+    request += addBoolToRequest("excludeRatedEnjoyment", removeRatedEnj);
+    request += addValueToRequest("minSubmissionCount", subLowCount, 0);
+    request += addValueToRequest("maxSubmissionCount", subHighCount, 0);
+    request += addValueToRequest("minEnjoymentCount", enjLowCount, 0);
+    request += addValueToRequest("maxEnjoymentCount", enjHighCount, 0);
+    request += addValueToRequest("minEnjoyment", enjLow, 0.0f);
+    request += addValueToRequest("maxEnjoyment", enjHigh, highestEnjoyment);
     request += addStringToRequest("sort", sort[sortOptionIndex]);
     request += addStringToRequest("sortDirection", sortDirection[sortDirectionIndex]);
     log::info("Search request: {}", request);
@@ -533,13 +530,23 @@ std::vector<int> GDDLSearchLayer::parseResponse(const std::string& response) {
     const auto maybeResponseJson = matjson::parse(response);
     if (maybeResponseJson.isOk()) {
         const matjson::Value& responseJson = maybeResponseJson.unwrap();
-        const int total = responseJson["total"].asInt().unwrap();
+        if (!responseJson.contains("levels")) {
+            // well, the json is probably wrong
+            const std::string error = responseJson["message"].asString().unwrapOr("Server returned invalid response - unknown error");
+            Notification::create(error, NotificationIcon::Error, 2)->show();
+            return results;
+        }
+        const int total = responseJson["total"].asInt().unwrapOr(-1);
+        if (total == -1) {
+            Notification::create("Server returned invalid response - no total results amount", NotificationIcon::Error, 2)->show();
+            return results;
+        }
         totalOnlineResults = std::max(totalOnlineResults, total); // so it never grabs 0 if a bad request is made
         matjson::Value levelList = responseJson["levels"];
         for (auto element: levelList.asArray().unwrap()) {
-            const int levelID = element["ID"].asInt().unwrap();
+            const int levelID = element["ID"].asInt().unwrapOr(0); // this way a missing ID won't matter
             if (levelID > 3) { // to avoid official demons
-                results.push_back(element["ID"].asInt().unwrap());
+                results.push_back(levelID);
                 if(!element["Rating"].isNull()) {
                     const float rating = element["Rating"].asDouble().unwrap();
                     RatingsManager::updateCacheFromSearch(levelID, rating);
@@ -596,7 +603,7 @@ int GDDLSearchLayer::getMaxPotentialPages() {
 }
 
 int GDDLSearchLayer::getOnlinePagesCount() {
-    const int chunk = time(nullptr) < Utils::API_SWITCH_TIME ? 50 : 25;
+    constexpr int chunk = 25;
     const int correction = totalOnlineResults % chunk == 0 ? 0 : 1;
     return totalOnlineResults / chunk + correction;
 }
@@ -604,7 +611,6 @@ int GDDLSearchLayer::getOnlinePagesCount() {
 GJSearchObject *GDDLSearchLayer::makeASearchObjectFrom(const int firstIndex, const int lastIndex) {
     std::string requestString;
     for (int i = firstIndex; i < lastIndex; i++) {
-        const int id = cachedResults[i];
         requestString += std::to_string(cachedResults[i]) + ',';
     }
     if (!requestString.empty()) {
@@ -634,9 +640,19 @@ void GDDLSearchLayer::appendFetchedResults(const std::string& response) {
 }
 
 std::pair<int, int> GDDLSearchLayer::getReadyRange(const int requestedPage) {
-    const int firstIndex = (requestedPage - 1) * 10;
+    const int firstIndex = requestedPage * 10;
     const int lastIndex = std::min(firstIndex + 10, static_cast<int>(cachedResults.size())); // last index + 1
     return {firstIndex, lastIndex};
+}
+
+void GDDLSearchLayer::hideAnyLoadingCircle() {
+    if (demonSplitLayer != nullptr) {
+        demonSplitLayer->hideLoadingCircle();
+        demonSplitLayer = nullptr;
+    }
+    if (searchLayer != nullptr) {
+        searchLayer->hideLoadingCircle();
+    }
 }
 
 void GDDLSearchLayer::handleSearchObject(GJSearchObject *searchObject, GDDLBrowserLayer* callbackObject,
@@ -645,13 +661,7 @@ void GDDLSearchLayer::handleSearchObject(GJSearchObject *searchObject, GDDLBrows
         callbackObject->handleSearchObject(searchObject, resultsCount);
     } else { // new search
         // remove any loading circles
-        if (demonSplitLayer != nullptr) {
-            demonSplitLayer->hideLoadingCircle();
-            demonSplitLayer = nullptr;
-        }
-        if (searchLayer != nullptr) {
-            searchLayer->hideLoadingCircle();
-        }
+        hideAnyLoadingCircle();
         // show the results
         const auto listLayer = LevelBrowserLayer::create(searchObject);
         const auto listLayerScene = CCScene::create();
@@ -664,25 +674,36 @@ void GDDLSearchLayer::handleSearchObject(GJSearchObject *searchObject, GDDLBrows
 void GDDLSearchLayer::prepareSearchListener() {
     searchListener.bind([] (web::WebTask::Event* e) {
             if (web::WebResponse* res = e->getValue()) {
-                const std::string response = res->string().unwrapOrDefault();
-                if (response.empty()) {
-                    Notification::create("Search failed - server error", NotificationIcon::Error, 2)->show();
-                } else {
-                    appendFetchedResults(response);
-                    auto [fst, snd] = getReadyRange(requestRequestedPage);
-                    if (snd - fst < 10 && onlinePagesFetched < getOnlinePagesCount()) {
-                        // recurse
-                        const std::string anotherRequest = formSearchRequest();
-                        auto req = web::WebRequest();
-                        req.header("User-Agent", Utils::getUserAgent());
-                        searchListener.setFilter(req.get(anotherRequest));
+                if (res->code() == 200) {
+                    const std::string response = res->string().unwrapOrDefault();
+                    if (response.empty()) {
+                        Notification::create("Search failed - received empty response", NotificationIcon::Error, 2)->show();
                     } else {
-                        GJSearchObject *searchObject = makeASearchObjectFrom(fst, snd);
-                        handleSearchObject(searchObject, searchCallbackObject, snd - fst);
+                        appendFetchedResults(response);
+                        auto [fst, snd] = getReadyRange(requestRequestedPage);
+                        if (snd - fst < 10 && onlinePagesFetched < getOnlinePagesCount()) {
+                            // recurse
+                            const std::string anotherRequest = formSearchRequest();
+                            auto req = web::WebRequest();
+                            req.header("User-Agent", Utils::getUserAgent());
+                            searchListener.setFilter(req.get(anotherRequest));
+                        } else {
+                            GJSearchObject *searchObject = makeASearchObjectFrom(fst, snd);
+                            handleSearchObject(searchObject, searchCallbackObject, snd - fst);
+                        }
                     }
+                } else {
+                    // not success!
+                    const auto jsonResponse = res->json().unwrapOr(matjson::Value());
+                    const std::string error = jsonResponse["message"].asString().unwrapOr("Search failed - unknown error");
+                    stopSearch();
+                    hideAnyLoadingCircle();
+                    Notification::create(error, NotificationIcon::Error, 2)->show();
                 }
             } else if (e->isCancelled()) {
-                Notification::create("Search failed - check your internet connection!", NotificationIcon::Error, 2)->show();
+                stopSearch();
+                hideAnyLoadingCircle();
+                Notification::create("Search failed - request cancelled", NotificationIcon::Error, 2)->show();
             }
         });
 }
@@ -935,7 +956,7 @@ void GDDLSearchLayer::onSearchClicked(CCObject *sender) {
     onlinePagesFetched = 0;
     searching = true;
     showLoadingCircle();
-    requestSearchPage(1, nullptr);
+    requestSearchPage(0, nullptr);
 }
 
 void GDDLSearchLayer::onResetClicked(CCObject *sender) { resetValues(); }
@@ -979,7 +1000,7 @@ void GDDLSearchLayer::onTierSearch(CCObject *sender) {
     onlinePagesFetched = 0;
     searching = true;
     showLoadingCircle();
-    requestSearchPage(1, nullptr);
+    requestSearchPage(0, nullptr);
 }
 
 void GDDLSearchLayer::setNumberFloatTextfield(const float value, CCTextInputNode *&textfield) {
@@ -1098,8 +1119,8 @@ void GDDLSearchLayer::saveSettings() {
 
 void GDDLSearchLayer::requestSearchPage(int requestedPage, GDDLBrowserLayer *callbackObject) {
     // check whether the cache already contains results for this query
-    if (requestedPage < 1) {
-        requestedPage = 1;
+    if (requestedPage < 0) {
+        requestedPage = 0;
     }
     if (!cachedResults.empty()) {
         const int maxPotentialPages = getMaxPotentialPages();
@@ -1147,7 +1168,7 @@ void GDDLSearchLayer::requestSearchFromDemonSplit(const int tier, GDDLDemonSplit
     prepareSearchListener();
     // save the layer to remove the loading circle later
     demonSplitLayer = layer;
-    requestSearchPage(1, nullptr);
+    requestSearchPage(0, nullptr);
 }
 
 int GDDLSearchLayer::getSearchResultsPageCount() { return getMaxPotentialPages(); }